Avoid names differing only by their case
The former functions `wignerd` and `wignerD` only differed by the case of
their last letter. This could potentially be a problem. So far, it only
resulted in an error of sphinx. However, this unpleasant naming (also
having a function that is not completely lowercase) is avoided with the
commit.

The function `wignerd` is now called `wignersmalld` and the function
`wignerD` takes the "free" name `wignerd`. The shorter name is taken for
the more general Wigner D-Matrix elements. Since the functions' argument
numbers differ by 2, it should also always throw an exception if they are
used with the former names in mind, which avoids using the wrong function
unintentionally.