Add support for Windows
In the past the only way to use ptsa on Windows was using WSL
(Windows Subsystem for Linux), now a different mode is possible,
using msys2. With msys2 it is possible to use the compiler gcc of
mingw-w64 and a matching python distribution to use ptsa.

Support for MSVC is not achieved with this commit, since it uses a
non-standard implementation of complex numbers. Simple workarounds
were not successful so far.